WIP: simplify/reorganize reconstruction plugin
